+++ CHANGELOG.rst
@@ -8,6 +8,10 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 Unreleased
 ==========
 
+Fixed
+-----
+
+- Albums and tracks without extras now show when queried.
 
 v0.8.2 (2022-02-03)
 ===================